+++ examples/client/generate_brfc/generate_brfc.go
@@ -1,13 +1,11 @@
 package main
 
 import (
-	"log"
-
 	"github.com/bitcoin-sv/go-paymail"
+	"log"
 )
 
 func main() {
-
 	// Start with a new BRFC specification
 	newBRFC := &paymail.BRFCSpec{
 		Author:  "MrZ",
@@ -19,5 +17,5 @@ func main() {
 	if err := newBRFC.Generate(); err != nil {
 		log.Fatalf("error generating BRFC id: %s", err.Error())
 	}
-	log.Printf("id generated: %s", newBRFC.ID)
+	log.Fatalf("id generated: %s", newBRFC.ID)
 }