+++ examples/client/get_srv_record/get_srv_record.go
@@ -8,7 +8,6 @@ import (
 )
 
 func main() {
-
 	// Load the client
 	client, err := paymail.NewClient()
 	if err != nil {
@@ -18,7 +17,7 @@ func main() {
 	// Get the SRV record
 	var srv *net.SRV
 	if srv, err = client.GetSRVRecord(paymail.DefaultServiceName, paymail.DefaultProtocol, "moneybutton.com"); err != nil {
-		log.Fatal("error getting SRV record: " + err.Error())
+		log.Fatalf("error getting SRV record: %s", err.Error())
 	}
-	log.Println("found SRV record:", srv)
+	log.Printf("found SRV record: %v", srv)
 }