+++ server/mock_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/bitcoin-sv/go-paymail"
+	"github.com/bitcoin-sv/go-paymail/spv"
 )
 
 // Mock implementation of a service provider
@@ -39,4 +40,15 @@ func (m *mockServiceProvider) RecordTransaction(_ context.Context,
 
 	// Record the tx into your datastore layer
 	return nil, nil
+}
+
+// VerifyMerkleRoots is a mock implementation of this interface
+func (m *mockServiceProvider) VerifyMerkleRoots(_ context.Context, _ []*spv.MerkleRootConfirmationRequestItem) error {
+
+	// Verify the merkle roots
+	return nil
+}
+
+func (m *mockServiceProvider) AddContact(ctx context.Context, requesterPaymail string, contact *paymail.PikeContactRequestPayload) error {
+	return nil
 }