+++ server/pki.go
@@ -1,41 +1,35 @@
 package server
 
 import (
+	"github.com/gin-gonic/gin"
 	"net/http"
 
 	"github.com/bitcoin-sv/go-paymail"
-	"github.com/julienschmidt/httprouter"
 )
 
 // showPKI will return the public key information for the corresponding paymail address
 //
 // Specs: http://bsvalias.org/03-public-key-infrastructure.html
-func (c *Configuration) showPKI(w http.ResponseWriter, req *http.Request, _ httprouter.Params) {
+func (c *Configuration) showPKI(context *gin.Context) {
+	incomingPaymail := context.Param(PaymailAddressParamName)
 
-	// Get the params & paymail address submitted via URL request
-	params := req.URL.Query()
-	incomingPaymail := params.Get("paymailAddress")
-
-	// Parse, sanitize and basic validation
 	alias, domain, address := paymail.SanitizePaymail(incomingPaymail)
 	if len(address) == 0 {
-		ErrorResponse(w, ErrorInvalidParameter, "invalid paymail: "+incomingPaymail, http.StatusBadRequest)
+		ErrorResponse(context, ErrorInvalidParameter, "domain unknown: "+domain, http.StatusBadRequest)
 		return
 	} else if !c.IsAllowedDomain(domain) {
-		ErrorResponse(w, ErrorUnknownDomain, "domain unknown: "+domain, http.StatusBadRequest)
+		ErrorResponse(context, ErrorUnknownDomain, "domain unknown: "+domain, http.StatusBadRequest)
 		return
 	}
 
-	// Create the metadata struct
-	md := CreateMetadata(req, alias, domain, "")
+	md := CreateMetadata(context.Request, alias, domain, "")
 
-	// Get from the data layer
-	foundPaymail, err := c.actions.GetPaymailByAlias(req.Context(), alias, domain, md)
+	foundPaymail, err := c.actions.GetPaymailByAlias(context.Request.Context(), alias, domain, md)
 	if err != nil {
-		ErrorResponse(w, ErrorFindingPaymail, err.Error(), http.StatusExpectationFailed)
+		ErrorResponse(context, ErrorFindingPaymail, err.Error(), http.StatusExpectationFailed)
 		return
 	} else if foundPaymail == nil {
-		ErrorResponse(w, ErrorPaymailNotFound, "paymail not found", http.StatusNotFound)
+		ErrorResponse(context, ErrorPaymailNotFound, "paymail not found: "+incomingPaymail, http.StatusBadRequest)
 		return
 	}
 
@@ -45,6 +39,5 @@ func (c *Configuration) showPKI(w http.ResponseWriter, req *http.Request, _ http
 		PubKey:   foundPaymail.PubKey,
 	}
 
-	// Set the response
-	writeJsonResponse(w, http.StatusOK, pkiPayload)
+	context.JSON(http.StatusOK, pkiPayload)
 }